requirements.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
wincertstore==0.2
django-cors-headers==3.9.0
django-widget-tweaks==1.4.8
Pillow==8.0.1